Add cristim/autospotting
AutoSpotting is a tool I wrote that automatically replaces on-demand instances in existing AutoScaling with spot instances, using the Attach/Detach AutoScaling APIs.

Each launched spot instance can be chosen from any cheaper available instance type available on the spot market(compared to the on-demand price of the original instance), as long as it provides at least as much capacity as the original instance.

The original instances may not even be available on the spot market, like for example t2 burstable instances.

It usually gives the same type, but it can often be larger (with more CPU cores, more memory, more instance store), and is configured identically as the original instance, same AZ, storage configuration, user_data, SGs, tags, etc.

See the README file in the repo for more information.