Add cavaliercoder/vpc-free - Find free IP address blocks in AWS EC2
Managing IPs can be difficult in any network if
not managed effectively from the start. This
tool is awesome because it creates a clear
picture of EC2 VPCs so that good decisions can
be made for building or refactoring networks.
It can also get you out of a pickle when IPs
are limited and you can't find a free block
when searching by hand.