Add Lambda API
Lambda API is a lightweight web framework for use with AWS Lambda and API Gateway. It has ZERO dependencies, is super lightweight (136KB), and was built specifically for Lambda's single run executions. It has a similar interface to ExpressJS, plus a powerful middleware system, built in logging, and binary support.